♻️ refactor(vectors): hoist shared rendering and shape logic onto AbstractVector
`Point` and `Tangent` carried byte-identical `aval`, `shape`, and `__pdoc__`
implementations, and the compact angle-bracket renderer existed three times
(`point.vectorform_pdoc`, `tangent._vectorform_pdoc`, and a `Coordinate`
variant that imported both of the former's helpers).

All of it depends only on `.chart` and `.data`, which every `AbstractVector`
has, so it now lives once in `base.py`:

- `AbstractVector.aval` / `.shape` are concrete (`shape` was abstract);
  `Coordinate` keeps its bundle-aware overrides.
- `AbstractVector.__pdoc__` covers both the vector form and the
  constructor-style form; `Coordinate` keeps its override.
- `vector_comps_unit_docs`, `vector_values_str`, and `vectorform_pdoc` are
  module-level in `base.py`. The unused `class_name` override parameter is
  dropped, and `vector_comps_unit_docs` loses a redundant branch: the
  mixed-unit formatter (`f"{c}[{unit}]" if unit else c`) already renders the
  all-unitless case.

No behaviour change; `repr`/`str` output is unchanged.
